Change src directory to use mediator pattern
